fix(core): Use Loader and Updater directly
Loader and Updater were loader as Composer packages. But it's impossible to use them in plugin scope. Because they share the same namespace in each plugin, if tow of these plugins are used on the same website PHP will pick the first which is found. That can cause a lot of issues, e.g. maybe second plugin requires larger Updater version.